Unify execution duration rendering in session header cells
Share the `Tag` + `ClockIcon` component between session header and trace
rows via a new `ExecutionDurationTag` and extract `normalizeDurationString`
to `DisplayUtils` so session totals like `34.00000000000002ms` render as
`34ms`.
